+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### 🎁 New Features
 
+* Added new `logData` option to `TrackService.track()` - allows applications to request that
+  key/value pairs provided within the `data` block of a track statement be logged along with the
+  standard output. Client-side support for this feature on a per-call basis added
+  in `hoist-react >= 57.1`, can also be defaulted within the `xhActivityTrackingConfig` app config.
 * Deprecated config `xhAppVersionCheckEnabled` in favor of object based `xhAppVersionCheck`. Apps
   will seamlessly migrate the existing value to this new config's `shouldUpdate` flag. The new
   config also supports a `shouldRequireRefresh` flag, which will force users to refresh their