+++ CHANGELOG.md
@@ -9,8 +9,8 @@
   standard output. Client-side support for this feature on a per-call basis added
   in `hoist-react >= 57.1`, can also be defaulted within the `xhActivityTrackingConfig` app config.
 * Deprecated config `xhAppVersionCheckEnabled` in favor of object based `xhAppVersionCheck`. Apps will
-  seamlessly migrate the existing value to this new config's `mode` flag. While backwards compatible
-  with older versions of hoist-core, the new `forceReload` mode requires `hoist-core >= v16.4.0`.
+  seamlessly migrate the existing value to this new config's `mode` flag.  This supports the new
+  `forceRefresh` mode introduced in hoist-react v58.
 
 ## 16.3.0 - 2023-06-20